+++ baybe/utils/interval.py
@@ -1,6 +1,7 @@
 """Utilities for handling intervals."""
 
 import sys
+import warnings
 from collections.abc import Iterable
 from functools import singledispatchmethod
 from typing import Any, Union
@@ -54,19 +55,45 @@ class Interval:
             )
 
     @property
+    def is_closed(self):
+        """Check whether the interval is closed."""
+        return np.isfinite(self.lower) and np.isfinite(self.upper)
+
+    @property
+    def is_half_open(self):
+        """Check whether the interval is half-open."""
+        return np.isfinite(self.lower) ^ np.isfinite(self.upper)
+
+    @property
+    def is_open(self):
+        """Check whether the interval is open."""
+        return (not np.isfinite(self.lower)) and (not np.isfinite(self.upper))
+
+    @property
     def is_finite(self):
         """Check whether the interval is finite."""
+        warnings.warn(
+            "The use of 'Interval.is_finite' is deprecated and will be disabled in "
+            "a future version. Use 'Interval.is_closed' instead.",
+            DeprecationWarning,
+        )
         return np.isfinite(self.lower) and np.isfinite(self.upper)
 
     @property
     def is_bounded(self):
         """Check whether the interval is bounded."""
+        warnings.warn(
+            "The use of 'Interval.is_bounded' is deprecated and will be disabled in "
+            "a future version. Use 'Interval.is_bounded' or 'Interval.is_half_open' "
+            "instead, depending on the situation.",
+            DeprecationWarning,
+        )
         return np.isfinite(self.lower) or np.isfinite(self.upper)
 
     @property
     def center(self):
-        """The center of the interval. Only applicable for finite intervals."""
-        if not self.is_finite:
+        """The center of the interval. Only applicable for closed intervals."""
+        if not self.is_closed:
             raise InfiniteIntervalError(
                 f"The interval {self} is infinite and thus has no center."
             )
@@ -91,7 +118,7 @@ class Interval:
         return Interval(*bounds)
 
     def to_tuple(self):
-        """Transfor the interval to a tuple."""
+        """Transform the interval to a tuple."""
         return self.lower, self.upper
 
     def to_ndarray(self):